More fixes

+++ aten/src/ATen/ParallelNative.cpp
@@ -222,8 +222,7 @@ void set_num_threads(int nthreads) {
     int stored_nthreads = num_intraop_threads.load();
     if (stored_nthreads <= 0) {
       // plus one because of master thread
-      // NOLINTNEXTLINE(cppcoreguidelines-narrowing-conversions,bugprone-narrowing-conversions)
-      stored_nthreads = _get_intraop_pool().size() + 1;
+      stored_nthreads = static_cast<int>(_get_intraop_pool().size() + 1);
     }
     if (stored_nthreads != nthreads) {
       TORCH_WARN(
@@ -251,8 +250,7 @@ int get_num_threads() {
     return intraop_default_num_threads();
   } else {
     TORCH_INTERNAL_ASSERT(nthreads == CONSUMED);
-    // NOLINTNEXTLINE(cppcoreguidelines-narrowing-conversions,bugprone-narrowing-conversions)
-    return _get_intraop_pool().size() + 1;
+    return static_cast<int>(_get_intraop_pool().size() + 1);
   }
 #else
   caffe2::PThreadPool* const pool = caffe2::pthreadpool();
